move forge update check to forge package & rename redirect
