+++ src/main/java/com/mitchej123/hodgepodge/mixins/early/forge/MixinForgeVersion_FixUpdateCheck.java
@@ -1,4 +1,4 @@
-package com.mitchej123.hodgepodge.mixins.early.minecraft;
+package com.mitchej123.hodgepodge.mixins.early.forge;
 
 import java.net.MalformedURLException;
 import java.net.URL;
@@ -15,7 +15,7 @@ public abstract class MixinForgeVersion_FixUpdateCheck {
      */
     // MC-Dev plugin can't handle inner classes, resulting in incorrect errors
     @Redirect(method = "run()V", at = @At(value = "NEW", target = "(Ljava/lang/String;)Ljava/net/URL;"), remap = false)
-    private URL redirectVersionCheck(String dud) throws MalformedURLException {
+    private URL hodgepodge$redirectVersionCheck(String dud) throws MalformedURLException {
         return new URL("https://files.minecraftforge.net/maven/net/minecraftforge/forge/promotions_slim.json");
     }
 }
